se crea nuevo metodo summaryReportsForAssignCases
